Create model_hosting.versions API.
